+++ packages/components/pages/f-mfa/src/components/Mfa.vue
@@ -211,13 +211,13 @@ export default {
         const isEmailValid = this.isPropertyValid(this.email?.toLowerCase(), 'email', EMAIL_RFC5322_REGEX);
         const isCodeValid = this.isPropertyValid(this.code, 'code', MFA_CODE_REGEX);
 
-        if (!isEmailValid || !isCodeValid) {
+        if (isEmailValid && isCodeValid) {
+            this.$log.info('MFA page loaded successfully');
+            this.$gtm.pushEvent(buildEvent(MFA_VISIBLE));
+        } else {
             this.showErrorPage = true;
             this.$log.warn('Error loading MFA page');
             this.$gtm.pushEvent(buildEvent(ERROR_VISIBLE));
-        } else {
-            this.$log.info('MFA page loaded successfully');
-            this.$gtm.pushEvent(buildEvent(MFA_VISIBLE));
         }
     },