+++ packages/components/pages/f-mfa/vue.config.js
@@ -22,7 +22,8 @@ module.exports = {
             '@justeat/f-button': '@justeat/f-button',
             '@justeat/f-card': '@justeat/f-card',
             '@justeat/f-card-with-content': '@justeat/f-card-with-content',
-            '@justeat/f-form-field': '@justeat/f-form-field'
+            '@justeat/f-form-field': '@justeat/f-form-field',
+            '@justeat/f-link': '@justeat/f-link'
         });
     },
     pluginOptions: {